+++ src/core/utils/html.js
@@ -1,6 +1,7 @@
 export function escapeHtml(str) {
     return str
         .replace(/&/g, '&amp;')
+        .replace(/"/g, '&quot;')
         .replace(/</g, '&lt;')
         .replace(/>/g, '&gt;');
 }